+++ src/main/java/com/supergt/racingdriver/response/MessageResponse.java
@@ -1,0 +1,53 @@
+package com.supergt.racingdriver.response;
+
+public class MessageResponse {
+    private String message;
+
+    public MessageResponse(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public class RacerCreateResponse {
+
+        private String message;
+
+        public RacerCreateResponse(String message) {
+            this.message = message;
+        }
+
+        public String getMessage() {
+            return message;
+        }
+    }
+
+    public class RacerUpdateResponse {
+
+        private String message;
+
+        public RacerUpdateResponse(String message) {
+            this.message = message;
+        }
+
+        public String getMessage() {
+            return message;
+        }
+    }
+
+    public class RacerDeleteResponse {
+
+        private String message;
+
+        public RacerDeleteResponse(String message) {
+            this.message = message;
+        }
+
+        public String getMessage() {
+            return message;
+        }
+    }
+}
+
