+++ src/main/java/com/supergt/racingdriver/response/RacerResponse.java
@@ -24,34 +24,4 @@ public class RacerResponse {
     public LocalDate getBirthday() {
         return birthday;
     }
-
-    public class Response {
-        private String message;
-
-        public Response(String message) {
-            this.message = message;
-        }
-
-        public String getMessage() {
-            return message;
-        }
-    }
-
-    public class RacerCreateResponse extends Response {
-        public RacerCreateResponse(String message) {
-            super(message);
-        }
-    }
-
-    public class RacerUpdateResponse extends Response {
-        public RacerUpdateResponse(String message) {
-            super(message);
-        }
-    }
-
-    public class RacerDeleteResponse extends Response {
-        public RacerDeleteResponse(String message) {
-            super(message);
-        }
-    }
 }
